+++ Ember/src/Ember/Util/PlatformUtils.h
@@ -4,6 +4,6 @@ namespace Ember {
 	class Time
 	{
 	public:
-		static float GetTime();
+		static Uint64 GetTime();
 	};
 }
